Add get_text_size function in utils.py
This commit introduces a new function, get_text_size, in the utils.py file. This function takes in text and font as parameters and returns the width and height of the text. The calculation of text dimensions is based on the font metrics and the bounding box of the text mask. This function will be useful in scenarios where we need to calculate the size of a text string for layout purposes.